[SPARK-29116][PYTHON][ML] Refactor py classes related to DecisionTree
